refactor(common): remove stdlib-replaceable shims
- Switch logging timestamps to Dates (drop date_time/time)
- Use UUIDs.uuid4 for naming/labels (drop uuid/device_random)
- Remove encoding/file helpers in favor of Base read + local checks
- Replace SmallRegistry log subject registry with Dict
- Move common error/log subject registration into error/logging
- Delete unused common modules and update tests

Tests: Reseau, AwsHTTP, HTTP
