+++ Project.toml
@@ -4,6 +4,8 @@ version = "1.1.1"
 
 [deps]
 Base64 = "2a0f44e3-6c83-55bd-87e4-b1978d98bd5f"
+Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
+UUIDs = "cf7118a7-6976-5b1a-9a39-7adc72f591a4"
 LibAwsCommon = "c6e421ba-b5f8-4792-a1c4-42948de3ed9d"
 LibAwsCal = "ef519ef6-af43-41a0-8ac4-eb81328190af"
 EnumX = "4e289a0a-7415-4d19-859d-a7e5c4648b56"
