+++ test/common_tests.jl
@@ -157,111 +157,3 @@ end
         @test true
     end
 end
-
-@testset "Home/temp env fallbacks" begin
-    keys = @static if Sys.iswindows()
-        ["USERPROFILE", "HOMEDRIVE", "HOMEPATH", "TMP", "TEMP"]
-    else
-        ["HOME", "TMPDIR", "TMP", "TEMP"]
-    end
-
-    old = Dict{String, Union{Nothing, String}}()
-    for k in keys
-        old[k] = get(ENV, k, nothing)
-    end
-
-    try
-        for k in keys
-            ENV[k] = ""
-        end
-
-        Reseau.reset_error()
-        home = Reseau.get_home_directory()
-        @test Reseau.string_len(home) == 0
-        @test Reseau.last_error() == Reseau.ERROR_GET_HOME_DIRECTORY_FAILED
-
-        tmp = Reseau.get_temp_directory()
-        @test !isempty(tmp)
-        @static if !Sys.iswindows()
-            @test tmp == "/tmp"
-        end
-    finally
-        for k in keys
-            v = old[k]
-            if v === nothing
-                pop!(ENV, k, nothing)
-            else
-                ENV[k] = v
-            end
-        end
-    end
-end
-
-@testset "UUID" begin
-    u = Ref{Reseau.uuid}()
-    @test Reseau.uuid_init(u) == Reseau.OP_SUCCESS
-    out = Ref(Reseau.ByteBuffer(Reseau.UUID_STR_LEN))
-    @test Reseau.uuid_to_str(u, out) == Reseau.OP_SUCCESS
-    @test out[].len == Reseau.UUID_STR_LEN - 1
-end
-
-@testset "UUID from string" begin
-    s = "123e4567-e89b-12d3-a456-426614174000"
-    u = Ref{Reseau.uuid}()
-    cursor = Ref(Reseau.ByteCursor(s))
-    @test Reseau.uuid_init_from_str(u, cursor) == Reseau.OP_SUCCESS
-    out = Ref(Reseau.ByteBuffer(Reseau.UUID_STR_LEN))
-    @test Reseau.uuid_to_str(u, out) == Reseau.OP_SUCCESS
-    @test String(Reseau.byte_cursor_from_buf(out[])) == s
-end
-
-@testset "Date time parsing" begin
-    dt = Ref{Reseau.date_time}()
-
-    rfc = "Sun, 06 Nov 1994 08:49:37 GMT"
-    rfc_cursor = Ref(Reseau.ByteCursor(rfc))
-    @test Reseau.date_time_init_from_str_cursor(dt, rfc_cursor, Reseau.DateFormat.RFC822) == Reseau.OP_SUCCESS
-    @test dt[].gmt_time.tm_year + 1900 == 1994
-    @test dt[].gmt_time.tm_mon + 1 == 11
-    @test dt[].gmt_time.tm_mday == 6
-    @test dt[].gmt_time.tm_hour == 8
-    @test dt[].gmt_time.tm_min == 49
-    @test dt[].gmt_time.tm_sec == 37
-
-    iso = "1994-11-06T08:49:37Z"
-    iso_cursor = Ref(Reseau.ByteCursor(iso))
-    @test Reseau.date_time_init_from_str_cursor(dt, iso_cursor, Reseau.DateFormat.ISO_8601) == Reseau.OP_SUCCESS
-    @test dt[].gmt_time.tm_year + 1900 == 1994
-    @test dt[].gmt_time.tm_mon + 1 == 11
-    @test dt[].gmt_time.tm_mday == 6
-    @test dt[].gmt_time.tm_hour == 8
-    @test dt[].gmt_time.tm_min == 49
-    @test dt[].gmt_time.tm_sec == 37
-
-    iso_basic = "19941106T084937Z"
-    iso_basic_cursor = Ref(Reseau.ByteCursor(iso_basic))
-    @test Reseau.date_time_init_from_str_cursor(dt, iso_basic_cursor, Reseau.DateFormat.ISO_8601_BASIC) == Reseau.OP_SUCCESS
-    @test dt[].gmt_time.tm_year + 1900 == 1994
-    @test dt[].gmt_time.tm_mon + 1 == 11
-    @test dt[].gmt_time.tm_mday == 6
-    @test dt[].gmt_time.tm_hour == 8
-    @test dt[].gmt_time.tm_min == 49
-    @test dt[].gmt_time.tm_sec == 37
-
-    bad_cursor = Ref(Reseau.ByteCursor("not a date"))
-    @test Reseau.date_time_init_from_str_cursor(dt, bad_cursor, Reseau.DateFormat.AUTO_DETECT) == Reseau.OP_ERR
-end
-
-@testset "byte buffer init from file" begin
-    mktemp() do path, io
-        write(io, "filedata")
-        close(io)
-
-        buf_ref = Ref{Reseau.ByteBuffer}()
-        @test Reseau.byte_buf_init_from_file(buf_ref, path) == Reseau.AWS_OP_SUCCESS
-        buf = buf_ref[]
-        @test buf.len == 8
-        @test String(Reseau.byte_cursor_from_buf(buf)) == "filedata"
-        @test buf.mem[Int(buf.len) + 1] == 0x00
-    end
-end
